requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==3.2.8
django-extensions==3.1.3
django-filter==21.1
django-light==0.1.0.post3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
drf-jwt==1.13.4
flake8==4.0.1
flake8-quotes==3.3.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==1.7.1
pytz==2021.3
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
